fix obvious bugs
