+++ scripts/api.py
@@ -108,8 +108,6 @@ class ApiHijack(api.Api):
         return self.controlnet_any2img(
             any2img_request=txt2img_request,
             original_callback=ApiHijack.text2imgapi,
-            p_class=StableDiffusionProcessingTxt2Img,
-            script_runner=scripts.scripts_txt2img,
             is_img2img=False,
         )
 
@@ -117,40 +115,16 @@ class ApiHijack(api.Api):
         return self.controlnet_any2img(
             any2img_request=img2img_request,
             original_callback=ApiHijack.img2imgapi,
-            p_class=StableDiffusionProcessingImg2Img,
-            script_runner=scripts.scripts_img2img,
             is_img2img=True,
         )
 
-    def controlnet_any2img(self, any2img_request, original_callback, p_class, script_runner, is_img2img):
+    def controlnet_any2img(self, any2img_request, original_callback, is_img2img):
         any2img_request = nest_deprecated_cn_fields(any2img_request)
-        script_runner = create_cn_script_runner(script_runner, any2img_request.controlnet_units, is_img2img)
+        any2img_request.alwayson_scripts.update('ControlNet', {'args': [is_img2img, False, *[vars(unit) for unit in any2img_request.controlnet_units]]})
         delattr(any2img_request, 'controlnet_units')
-        with self.queue_lock:
-            self_copy = copy.copy(self)
-            self_copy.queue_lock = contextlib.nullcontext()
-            with OverrideInit(p_class, scripts=script_runner):
-                return original_callback(self_copy, any2img_request)
+        return original_callback(self, any2img_request)
 
 api.Api = ApiHijack
-
-class OverrideInit:
-    def __init__(self, cls, **kwargs):
-        self.cls = cls
-        self.kwargs = kwargs
-        self.original_init = None
-
-    def __enter__(self):
-        def init_hijack(p, *args, **kwargs):
-            self.original_init(p, *args, **kwargs)
-            for k, v in self.kwargs.items():
-                setattr(p, k, v)
-
-        self.original_init = self.cls.__init__
-        self.cls.__init__ = init_hijack
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        self.cls.__init__ = self.original_init
 
 def nest_deprecated_cn_fields(any2img_request):
     deprecated_cn_fields = {k: v for k, v in vars(any2img_request).items()
@@ -174,43 +148,6 @@ def nest_deprecated_cn_fields(any2img_request):
 
     any2img_request.controlnet_units.insert(0, ControlNetUnitRequest(**deprecated_cn_fields))
     return any2img_request
-
-def create_cn_script_runner(script_runner: scripts.ScriptRunner, control_unit_requests: List[ControlNetUnitRequest], is_img2img: bool):
-    if not script_runner.scripts:
-        script_runner.initialize_scripts(False)
-        ui.create_ui()
-
-    cn_script = external_code.find_cn_script(script_runner)
-    cn_script_runner = copy.copy(script_runner)
-    cn_script_runner.alwayson_scripts = [cn_script]
-    cn_script_args = [None] * cn_script.args_from
-    cn_units = [to_api_cn_unit(control_unit_request) for control_unit_request in control_unit_requests]
-    external_code.update_cn_script_in_place(
-        script_runner=cn_script_runner,
-        script_args=cn_script_args,
-        cn_units=cn_units,
-        is_img2img=is_img2img,
-    )
-
-    def make_script_runner_f_hijack(fixed_original_f):
-        def script_runner_f_hijack(p, *args, **kwargs):
-            original_script_args = p.script_args
-            try:
-                p.script_args = cn_script_args
-                fixed_original_f(p, *args, **kwargs)
-            finally:
-                p.script_args = original_script_args
-
-        return script_runner_f_hijack
-
-    for k in ('process', 'process_batch', 'postprocess', 'postprocess_batch', 'postprocess_image'):
-        original_f = getattr(cn_script_runner, k, None)
-        if original_f is None:
-            continue
-
-        setattr(cn_script_runner, k, make_script_runner_f_hijack(original_f))
-
-    return cn_script_runner
 
 def to_api_cn_unit(unit_request: ControlNetUnitRequest) -> external_code.ControlNetUnit:
     input_image = to_base64_nparray(unit_request.input_image) if unit_request.input_image else None
